requirements.txt
tensorflow>=2.15.0
Pillow>=10.1.0
opencv-python-headless>=4.8.1.78
numpy>=1.26.0
matplotlib>=3.8.0